Extract AnonymisableConfig to Anony::DSL
Using a separate class makes it easier to distinguish the options
available in the `anonymise` block from the other mechanics of
anonymisation.
